rearrange Rust routines into separate modules. Update examples, Python docstrings and Readme.
