+++ mannrs/__init__.py
@@ -1,4 +1,4 @@
-from .Mann import Stencil, ForgetfulStencil, save_box, load_mann_binary, mann_spectra
+from .Mann import Stencil, save_box, load_mann_binary, mann_spectra
 from .Constrained import ConstrainedStencil, Constraint
